requirements.txt
flask
psycopg2-binary
flask-login
flask-bcrypt
qrcode
reportlab
gunicorn
